+++ src/simulator/chip/builtins/sequential/ram.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
 import { Bus, ClockedChip, Pin } from "../../chip";
 import { assert } from "@davidsouther/jiffies/lib/esm/assert";
-import { Memory as MemoryComponent } from "../../../../components/memory";
+import { Memory as MemoryComponent } from "../../../../components/chips/memory";
 import { Memory, Memory as MemoryChip } from "../../../cpu/memory";
 
 export class RAM extends ClockedChip {
